Move some structs to utils
