+++ pkg/lib/authenticationflow/declarative/utils_account_linking.go
@@ -124,3 +124,38 @@ func linkByOAuthIncomingOAuthSpec(
 
 	return oauthConfig, conflicts, nil
 }
+
+type CreateIdentityRequest struct {
+	Type model.IdentityType
+
+	LoginID *CreateIdentityRequestLoginID
+	OAuth   *CreateIdentityRequestOAuth
+}
+
+type CreateIdentityRequestOAuth struct {
+	Alias string
+	Spec  *identity.Spec
+}
+
+type CreateIdentityRequestLoginID struct {
+	Spec *identity.Spec
+}
+
+func NewCreateOAuthIdentityRequest(alias string, spec *identity.Spec) *CreateIdentityRequest {
+	return &CreateIdentityRequest{
+		Type: model.IdentityTypeOAuth,
+		OAuth: &CreateIdentityRequestOAuth{
+			Alias: alias,
+			Spec:  spec,
+		},
+	}
+}
+
+func NewCreateLoginIDIdentityRequest(spec *identity.Spec) *CreateIdentityRequest {
+	return &CreateIdentityRequest{
+		Type: model.IdentityTypeLoginID,
+		LoginID: &CreateIdentityRequestLoginID{
+			Spec: spec,
+		},
+	}
+}
